vendor: drop inclusion of vendor-def stuff
why remove? basicially most of the stuffs are defined in their own repo like powerhals and stuff and for display libs that are not required even and  the required one are their in the repective device hals this commit will also include removing of repos like Snapdragoncamera ,Mms,QTIRecorder they are just legacy apps no one uses it. this is also help us to track less repos also so better to remove it. see ya!!